Full synf from the web.  working on sending back the link data WARNING!!!!! FIX THIS SOON!!!!
****
When adding a module in the app, you can only create encounters which are then not added to the master encounters table.  THis is different than in teh web, which will lead to object-replication. To fix, mirror the web and add an intermediary activity to select an existing encounter or create a new one.  Note that this might lead to a LOT of encounters to choose from, but oh well.

+++ app/src/main/java/net/kalinec/dndencounters/sync/SyncPayload.java
@@ -3,9 +3,9 @@ package net.kalinec.dndencounters.sync;
 public class SyncPayload
 {
 	public String uuid;
-	public int dbId;
+	public long dbId;
 
-	public SyncPayload(String uuid, int dbId)
+	public SyncPayload(String uuid, long dbId)
 	{
 		this.uuid = uuid;
 		this.dbId = dbId;
